render period cron expression in custom mode
